refactor: have schema in root for easier discoverability
